+++ src/components/harp-cell/harp-cell.tsx
@@ -22,7 +22,9 @@ export type YXCoord = [Coord, Coord]
 
 export type HarpCellStyles = {
   readonly cell: ViewStyle
-  readonly contentsWrapper: ViewStyle
+  readonly naturalContentsWrapper: ViewStyle
+  readonly sharpContentsWrapper: ViewStyle
+  readonly flatContentsWrapper: ViewStyle
   readonly note: TextStyle
   readonly modifier: TextStyle
 }
@@ -54,16 +56,38 @@ export const HarpCell = ({ yxCoord }: HarpCellProps): React.ReactElement => {
     setPozitionRoot(thisPitchId)
   }
 
+  const contentFragment =
+    displayValue.length === 2 ? (
+      <>
+        <View style={styles.sharpContentsWrapper}>
+          <Text style={styles.note}>{displayValue[0][0]}</Text>
+        </View>
+        <View style={styles.sharpContentsWrapper}>
+          <Text style={styles.modifier}>{displayValue[0][1]}</Text>
+        </View>
+        <View style={styles.flatContentsWrapper}>
+          <Text style={styles.note}>{displayValue[1][0]}</Text>
+        </View>
+        <View style={styles.flatContentsWrapper}>
+          <Text style={styles.modifier}>{displayValue[1][1]}</Text>
+        </View>
+      </>
+    ) : (
+      <>
+        <View style={styles.naturalContentsWrapper}>
+          <Text style={styles.note}>{displayValue[0][0]}</Text>
+        </View>
+        <View style={styles.naturalContentsWrapper}>
+          <Text style={styles.modifier}>{displayValue[0][1]}</Text>
+        </View>
+      </>
+    )
+
   const accessibleContent = (
     <LongPressGestureHandler onHandlerStateChange={handleLongPressStateChange}>
       <TapGestureHandler onHandlerStateChange={handleTapStateChange}>
         <View accessible={true} accessibilityRole="button" style={styles.cell}>
-          <View style={styles.contentsWrapper}>
-            <Text style={styles.note}>{displayValue[0]}</Text>
-          </View>
-          <View style={styles.contentsWrapper}>
-            <Text style={styles.modifier}>{displayValue[1]}</Text>
-          </View>
+          {contentFragment}
         </View>
       </TapGestureHandler>
     </LongPressGestureHandler>